requirements.txt
alembic==1.8.1
awscrt==0.14.0
bcrypt==4.0.1
blinker==1.5
boto3==1.26.8
botocore==1.29.8
cachetools==5.2.0
certifi==2022.9.24
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-DebugToolbar==0.10.1
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
greenlet==2.0.1
idna==2.6
importlib-metadata==5.0.0
importlib-resources==5.10.0
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==1.0.1
Mako==1.2.3
MarkupSafe==1.1.1
oauthlib==3.2.2
Pillow==9.3.0
psycopg2-binary==2.9.3
PyJWT==1.5.3
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.0
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.3.13
urllib3==1.26.12
Werkzeug==0.16.1
zipp==3.10.0
